Refactor: solemap 패키지명 수정 (#23)
