+++ src/main/java/com/ilta/solepli/domain/solemap/controller/SolemapController.java
@@ -1,4 +1,4 @@
-package com.ilta.solepli.domain.solmap.controller;
+package com.ilta.solepli.domain.solemap.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,8 +10,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 
-import com.ilta.solepli.domain.solmap.dto.ViewportMapMarkerResponse;
-import com.ilta.solepli.domain.solmap.service.SolemapService;
+import com.ilta.solepli.domain.solemap.dto.ViewportMapMarkerResponse;
+import com.ilta.solepli.domain.solemap.service.SolemapService;
 import com.ilta.solepli.global.response.SuccessResponse;
 
 @RestController
