+++ src/main/java/com/ilta/solepli/domain/solemap/dto/ViewportMapMarkerDetail.java
@@ -1,4 +1,4 @@
-package com.ilta.solepli.domain.solmap.dto;
+package com.ilta.solepli.domain.solemap.dto;
 
 import lombok.Builder;
