+++ src/main/java/com/ilta/solepli/domain/solemap/dto/ViewportMapMarkerResponse.java
@@ -1,4 +1,4 @@
-package com.ilta.solepli.domain.solmap.dto;
+package com.ilta.solepli.domain.solemap.dto;
 
 import java.util.List;
